+++ src/geophires_x/EconomicsSam.py
@@ -3,7 +3,6 @@ from __future__ import annotations
 import json
 import logging
 import os
-from dataclasses import dataclass, field
 from functools import lru_cache
 from math import isnan
 from pathlib import Path
@@ -29,26 +28,21 @@ from pint.facets.plain import PlainQuantity
 from tabulate import tabulate
 
 from geophires_x import Model as Model
+from geophires_x.EconomicsSamCalculations import (
+    SamEconomicsCalculations,
+    CapacityPaymentRevenueSource,
+    ROYALTIES_OPEX_CASH_FLOW_LINE_ITEM_KEY,
+    _after_tax_net_cash_flow_all_years,
+    _net_cash_flow_all_years,
+)
 from geophires_x.EconomicsSamCashFlow import (
     _calculate_sam_economics_cash_flow_operational_years,
-    _SAM_CASH_FLOW_NAN_STR,
 )
 from geophires_x.EconomicsUtils import (
     BuildPricingModel,
-    wacc_output_parameter,
-    nominal_discount_rate_parameter,
-    after_tax_irr_parameter,
-    moic_parameter,
-    project_vir_parameter,
-    project_payback_period_parameter,
-    total_capex_parameter_output_parameter,
-    royalty_cost_output_parameter,
-    overnight_capital_cost_output_parameter,
     _SAM_EM_MOIC_RETURNS_TAX_QUALIFIER,
-    investment_tax_credit_output_parameter,
 )
 from geophires_x.EconomicsSamPreRevenue import (
-    _AFTER_TAX_NET_CASH_FLOW_ROW_NAME,
     PreRevenueCostsAndCashflow,
     calculate_pre_revenue_costs_and_cashflow,
     adjust_phased_schedule_to_new_length,
@@ -56,543 +50,9 @@ from geophires_x.EconomicsSamPreRevenue import (
 from geophires_x.GeoPHIRESUtils import is_float, is_int, sig_figs, quantity
 from geophires_x.OptionList import EconomicModel, EndUseOptions
 from geophires_x.Parameter import Parameter, OutputParameter, floatParameter, listParameter
-from geophires_x.Units import convertible_unit, EnergyCostUnit, CurrencyUnit, Units
+from geophires_x.Units import convertible_unit
 
 _log = logging.getLogger(__name__)
-
-ROYALTIES_OPEX_CASH_FLOW_LINE_ITEM_KEY = 'O&M production-based expense ($)'
-
-
-@dataclass
-class CapacityPaymentRevenueSource:
-    name: str
-
-    revenue_usd: list[float]
-
-    amount_provided_label: str | None = None
-    amount_provided: list[float] | None = None
-
-    price_label: str | None = None
-    price: list[float] | None = None
-
-
-@dataclass
-class SamEconomicsCalculations:
-    _sam_cash_flow_profile_operational_years: list[list[Any]]
-    """
-    Operational cash flow profile from SAM financial engine
-    """
-
-    pre_revenue_costs_and_cash_flow: PreRevenueCostsAndCashflow
-
-    lcoe_nominal: OutputParameter = field(
-        default_factory=lambda: OutputParameter(
-            UnitType=Units.ENERGYCOST,
-            CurrentUnits=EnergyCostUnit.CENTSSPERKWH,
-        )
-    )
-
-    overnight_capital_cost: OutputParameter = field(default_factory=overnight_capital_cost_output_parameter)
-
-    capex: OutputParameter = field(default_factory=total_capex_parameter_output_parameter)
-
-    _royalties_rate_schedule: list[float] | None = None
-    royalties_opex: OutputParameter = field(default_factory=royalty_cost_output_parameter)
-
-    project_npv: OutputParameter = field(
-        default_factory=lambda: OutputParameter(
-            UnitType=Units.CURRENCY,
-            CurrentUnits=CurrencyUnit.MDOLLARS,
-        )
-    )
-
-    after_tax_irr: OutputParameter = field(default_factory=after_tax_irr_parameter)
-    nominal_discount_rate: OutputParameter = field(default_factory=nominal_discount_rate_parameter)
-    wacc: OutputParameter = field(default_factory=wacc_output_parameter)
-    moic: OutputParameter = field(default_factory=moic_parameter)
-    project_vir: OutputParameter = field(default_factory=project_vir_parameter)
-
-    project_payback_period: OutputParameter = field(default_factory=project_payback_period_parameter)
-
-    investment_tax_credit: OutputParameter = field(default_factory=investment_tax_credit_output_parameter)
-
-    capacity_payment_revenue_sources: list[CapacityPaymentRevenueSource] = field(default_factory=list)
-
-    @property
-    def _pre_revenue_years_count(self) -> int:
-        return len(
-            self.pre_revenue_costs_and_cash_flow.pre_revenue_cash_flow_profile_dict[_AFTER_TAX_NET_CASH_FLOW_ROW_NAME]
-        )
-
-    @property
-    def sam_cash_flow_profile(self) -> list[list[Any]]:
-        ret: list[list[Any]] = self._sam_cash_flow_profile_operational_years.copy()
-        col_count = len(self._sam_cash_flow_profile_operational_years[0])
-
-        # TODO support/insert calendar year line item https://github.com/NREL/GEOPHIRES-X/issues/439
-
-        pre_revenue_years_to_insert = self._pre_revenue_years_count - 1
-
-        construction_rows: list[list[Any]] = [
-            ['CONSTRUCTION'] + [''] * (len(self._sam_cash_flow_profile_operational_years[0]) - 1)
-        ]
-
-        for row_index in range(len(self._sam_cash_flow_profile_operational_years)):
-            pre_revenue_row_content = [''] * pre_revenue_years_to_insert
-            insert_index = 1
-
-            if row_index == 0:
-                for pre_revenue_year in range(pre_revenue_years_to_insert):
-                    negative_year_index: int = self._pre_revenue_years_count - 1 - pre_revenue_year
-                    pre_revenue_row_content[pre_revenue_year] = f'Year -{negative_year_index}'
-
-                for _, row_ in enumerate(self.pre_revenue_costs_and_cash_flow.pre_revenue_cash_flow_profile):
-                    pre_revenue_row = row_.copy()
-                    pre_revenue_row.extend([''] * (col_count - len(pre_revenue_row)))
-                    construction_rows.append(pre_revenue_row)
-
-            #  TODO zero-vectors for non-construction years e.g. Debt principal payment ($)
-
-            adjusted_row = [ret[row_index][0]] + pre_revenue_row_content + ret[row_index][insert_index:]
-            ret[row_index] = adjusted_row
-
-        construction_rows.append([''] * len(self._sam_cash_flow_profile_operational_years[0]))
-        for construction_row in reversed(construction_rows):
-            ret.insert(1, construction_row)
-
-        def _get_row_index(row_name_: str) -> list[Any]:
-            return [it[0] for it in ret].index(row_name_)
-
-        def _get_row(row_name__: str) -> list[Any]:
-            for r in ret:
-                if r[0] == row_name__:
-                    return r[1:]
-
-            raise ValueError(f'Could not find row with name {row_name__}')
-
-        after_tax_cash_flow: list[float] = (
-            self.pre_revenue_costs_and_cash_flow.pre_revenue_cash_flow_profile_dict[_AFTER_TAX_NET_CASH_FLOW_ROW_NAME]
-            + _get_row('Total after-tax returns ($)')[self._pre_revenue_years_count :]
-        )
-        after_tax_cash_flow = [float(it) for it in after_tax_cash_flow if is_float(it)]
-        irr_row_name = 'After-tax cumulative IRR (%)'
-        ret.insert(
-            _get_row_index(irr_row_name), ['After-tax net cash flow ($)', *[int(it) for it in after_tax_cash_flow]]
-        )
-
-        npv_usd = []
-        irr_pct = []
-        for year in range(len(after_tax_cash_flow)):
-            npv_usd.append(
-                round(
-                    npf.npv(
-                        self.nominal_discount_rate.quantity().to('dimensionless').magnitude,
-                        after_tax_cash_flow[: year + 1],
-                    )
-                )
-            )
-
-            year_irr = npf.irr(after_tax_cash_flow[: year + 1]) * 100.0
-            irr_pct.append(year_irr if not isnan(year_irr) else _SAM_CASH_FLOW_NAN_STR)
-
-        ret[_get_row_index('After-tax cumulative NPV ($)')] = ['After-tax cumulative NPV ($)'] + npv_usd
-        ret[_get_row_index('After-tax cumulative IRR (%)')] = [irr_row_name] + irr_pct
-
-        if self._royalties_rate_schedule is not None:
-            ret = self._insert_royalties_rate_schedule(ret)
-
-        ret = self._insert_capacity_payment_line_items(ret)
-
-        ret = self._insert_calculated_levelized_metrics_line_items(ret)
-
-        return ret
-
-    def _insert_royalties_rate_schedule(self, cf_ret: list[list[Any]]) -> list[list[Any]]:
-        ret = cf_ret.copy()
-
-        def _get_row_index(row_name_: str) -> list[Any]:
-            return [it[0] for it in ret].index(row_name_)
-
-        ret.insert(
-            _get_row_index(ROYALTIES_OPEX_CASH_FLOW_LINE_ITEM_KEY),
-            [
-                *['Royalty rate (%)'],
-                *([''] * (self._pre_revenue_years_count)),
-                *[
-                    quantity(it, 'dimensionless').to(convertible_unit('percent')).magnitude
-                    for it in self._royalties_rate_schedule
-                ],
-            ],
-        )
-
-        return ret
-
-    # noinspection PyMethodMayBeStatic
-    def _insert_capacity_payment_line_items(self, cf_ret: list[list[Any]]) -> list[list[Any]]:
-        if len(self.capacity_payment_revenue_sources) == 0:
-            return cf_ret
-
-        ret: list[list[Any]] = cf_ret.copy()
-
-        def _get_row_index(row_name_: str) -> list[Any]:
-            return [it[0] for it in ret].index(row_name_)
-
-        def _insert_row_before(before_row_name: str, row_name: str, row_content: list[Any]) -> None:
-            ret.insert(
-                _get_row_index(before_row_name),
-                [row_name, *row_content],
-            )
-
-        def _insert_blank_line_before(before_row_name: str) -> None:
-            _insert_row_before(before_row_name, '', ['' for _it in ret[_get_row_index(before_row_name)]][1:])
-
-        revenue_row_name = 'REVENUE'
-        capacity_payment_revenue_row_name = 'Capacity payment revenue ($)'
-
-        _insert_blank_line_before('Salvage value ($)')
-        _insert_blank_line_before(capacity_payment_revenue_row_name)
-
-        def _for_operational_years(_row: list[Any]) -> list[Any]:
-            return [*([''] * (self._pre_revenue_years_count - 1)), 0, *_row]
-
-        for capacity_payment_revenue_source in self.capacity_payment_revenue_sources:
-            if capacity_payment_revenue_source.amount_provided_label is not None:
-                _insert_row_before(
-                    revenue_row_name,
-                    capacity_payment_revenue_source.amount_provided_label,
-                    _for_operational_years(capacity_payment_revenue_source.amount_provided),
-                )
-                _insert_blank_line_before(revenue_row_name)
-
-            revenue_row_name = f'{capacity_payment_revenue_source.name} revenue ($)'
-            _insert_row_before(
-                capacity_payment_revenue_row_name,
-                revenue_row_name,
-                _for_operational_years(capacity_payment_revenue_source.revenue_usd),
-            )
-
-            if capacity_payment_revenue_source.price_label is not None:
-                _insert_row_before(
-                    revenue_row_name,
-                    capacity_payment_revenue_source.price_label.replace('USD', '$'),
-                    capacity_payment_revenue_source.price,
-                )
-
-        return ret
-
-    # noinspection DuplicatedCode
-    def _insert_calculated_levelized_metrics_line_items(self, cf_ret: list[list[Any]]) -> list[list[Any]]:
-        ret = cf_ret.copy()
-
-        __row_names: list[str] = [it[0] for it in ret]
-
-        def _get_row_index(row_name_: str, raise_exception_if_not_present: bool = True) -> int:
-            try:
-                return __row_names.index(row_name_)
-            except ValueError as ve:
-                if raise_exception_if_not_present:
-                    raise ve
-                else:
-                    return -1
-
-        def _get_row_indexes(row_name_: str, after_row_name: str | None = None) -> list[int]:
-            after_criteria_met: bool = True if after_row_name is None else False
-            indexes = []
-            for idx, _row_name_ in enumerate(__row_names):
-                if _row_name_ == after_row_name:
-                    after_criteria_met = True
-
-                if _row_name_ == row_name_ and after_criteria_met:
-                    indexes.append(idx)
-
-            return indexes
-
-        def _get_row_index_after(row_name_: str, after_row_name: str) -> int:
-            return _get_row_indexes(row_name_, after_row_name=after_row_name)[0]
-
-        after_tax_lcoe_and_ppa_price_header_row_title = 'AFTER-TAX LCOE AND PPA PRICE'
-
-        # Backfill annual costs
-        annual_costs_usd_row_name = 'Annual costs ($)'
-        annual_costs = cf_ret[_get_row_index(annual_costs_usd_row_name)].copy()
-        after_tax_net_cash_flow_usd = cf_ret[_get_row_index('After-tax net cash flow ($)')]
-
-        annual_costs_backfilled = [
-            *after_tax_net_cash_flow_usd[1 : (self._pre_revenue_years_count + 1)],
-            *annual_costs[(self._pre_revenue_years_count + 1) :],
-        ]
-
-        ret[_get_row_index(annual_costs_usd_row_name)][1:] = annual_costs_backfilled
-
-        ppa_revenue_row_name = 'PPA revenue ($)'
-        ppa_revenue_row_index = _get_row_index_after(
-            ppa_revenue_row_name, after_tax_lcoe_and_ppa_price_header_row_title
-        )
-        year_0_ppa_revenue: float = ret[ppa_revenue_row_index][self._pre_revenue_years_count]
-        if year_0_ppa_revenue != 0.0:
-            # Shouldn't happen
-            _log.warning(f'PPA revenue in Year 0 ({year_0_ppa_revenue}) is not zero, this is unexpected.')
-
-        ret[ppa_revenue_row_index][1 : self._pre_revenue_years_count] = [year_0_ppa_revenue] * (
-            self._pre_revenue_years_count - 1
-        )
-
-        electricity_to_grid_kwh_row_name = 'Electricity to grid (kWh)'
-        electricity_to_grid = cf_ret[_get_row_index(electricity_to_grid_kwh_row_name)].copy()
-        electricity_to_grid_backfilled = [
-            0 if it == '' else (int(it) if is_int(it) else it) for it in electricity_to_grid[1:]
-        ]
-
-        electricity_to_grid_kwh_row_index = _get_row_index_after(
-            electricity_to_grid_kwh_row_name, after_tax_lcoe_and_ppa_price_header_row_title
-        )
-        ret[electricity_to_grid_kwh_row_index][1:] = electricity_to_grid_backfilled
-
-        pv_of_annual_costs_backfilled_row_name = 'Present value of annual costs ($)'
-
-        # Backfill PV of annual costs
-        annual_costs_backfilled_pv_processed = annual_costs_backfilled.copy()
-        pv_of_annual_costs_backfilled = []
-        for year in range(self._pre_revenue_years_count):
-            pv_at_year = abs(
-                round(
-                    npf.npv(
-                        self.nominal_discount_rate.quantity().to('dimensionless').magnitude,
-                        annual_costs_backfilled_pv_processed,
-                    )
-                )
-            )
-
-            pv_of_annual_costs_backfilled.append(pv_at_year)
-
-            cost_at_year = annual_costs_backfilled_pv_processed.pop(0)
-            annual_costs_backfilled_pv_processed[0] = annual_costs_backfilled_pv_processed[0] + cost_at_year
-
-        pv_of_annual_costs_backfilled_row = [
-            *[pv_of_annual_costs_backfilled_row_name],
-            *pv_of_annual_costs_backfilled,
-        ]
-
-        pv_of_annual_costs_row_name = 'Present value of annual costs ($)'
-        pv_of_annual_costs_row_index = _get_row_index(pv_of_annual_costs_row_name)
-        ret[pv_of_annual_costs_row_index][1:] = [
-            pv_of_annual_costs_backfilled[0],
-            *([''] * (self._pre_revenue_years_count - 1)),
-        ]
-
-        # Backfill PV of electricity to grid
-        electricity_to_grid_backfilled_pv_processed = electricity_to_grid_backfilled.copy()
-        pv_of_electricity_to_grid_backfilled_kwh = []
-        for year in range(self._pre_revenue_years_count):
-            pv_at_year = abs(
-                round(
-                    npf.npv(
-                        self.nominal_discount_rate.quantity().to('dimensionless').magnitude,
-                        electricity_to_grid_backfilled_pv_processed,
-                    )
-                )
-            )
-
-            pv_of_electricity_to_grid_backfilled_kwh.append(pv_at_year)
-
-            electricity_to_grid_at_year = electricity_to_grid_backfilled_pv_processed.pop(0)
-            electricity_to_grid_backfilled_pv_processed[0] = (
-                electricity_to_grid_backfilled_pv_processed[0] + electricity_to_grid_at_year
-            )
-
-        pv_of_annual_energy_row_name = 'Present value of annual energy nominal (kWh)'
-        for pv_of_annual_energy_row_index in _get_row_indexes(pv_of_annual_energy_row_name):
-            ret[pv_of_annual_energy_row_index][1:] = [
-                pv_of_electricity_to_grid_backfilled_kwh[0],
-                *([''] * (self._pre_revenue_years_count - 1)),
-            ]
-
-        def backfill_lcoe_nominal() -> None:
-            pv_of_electricity_to_grid_backfilled_row_kwh = pv_of_electricity_to_grid_backfilled_kwh
-            pv_of_annual_costs_backfilled_row_values_usd = pv_of_annual_costs_backfilled_row[
-                1 if isinstance(pv_of_annual_costs_backfilled_row[0], str) else 0 :
-            ]
-
-            lcoe_nominal_backfilled = []
-            for _year in range(len(pv_of_annual_costs_backfilled_row_values_usd)):
-                entry: float | str = 'NaN'
-                if pv_of_electricity_to_grid_backfilled_row_kwh[_year] != 0:
-                    entry = (
-                        pv_of_annual_costs_backfilled_row_values_usd[_year]
-                        * 100
-                        / pv_of_electricity_to_grid_backfilled_row_kwh[_year]
-                    )
-
-                lcoe_nominal_backfilled.append(entry)
-
-            lcoe_nominal_row_name = 'LCOE Levelized cost of energy nominal (cents/kWh)'
-            lcoe_nominal_row_index = _get_row_index(lcoe_nominal_row_name)
-
-            lcoe_nominal_backfilled_entry = lcoe_nominal_backfilled[0]
-            if isinstance(lcoe_nominal_backfilled_entry, float):
-                lcoe_nominal_backfilled_entry = round(lcoe_nominal_backfilled_entry, 2)
-
-            ret[lcoe_nominal_row_index][1:] = [
-                lcoe_nominal_backfilled_entry,
-                *([None] * (self._pre_revenue_years_count - 1)),
-            ]
-
-        backfill_lcoe_nominal()
-
-        def backfill_lppa_metrics() -> None:
-            pv_of_ppa_revenue_row_index = _get_row_index_after(
-                'Present value of PPA revenue ($)', after_tax_lcoe_and_ppa_price_header_row_title
-            )
-            first_year_pv_of_ppa_revenue_usd = round(
-                npf.npv(
-                    self.nominal_discount_rate.quantity().to('dimensionless').magnitude,
-                    ret[ppa_revenue_row_index][1:],
-                )
-            )
-            ret[pv_of_ppa_revenue_row_index][1:] = [
-                first_year_pv_of_ppa_revenue_usd,
-                *([None] * (self._pre_revenue_years_count - 1)),
-            ]
-
-            ppa_price_row_index = _get_row_index('PPA price (cents/kWh)')
-            year_0_ppa_price: float = ret[ppa_price_row_index][self._pre_revenue_years_count]
-            if year_0_ppa_price != 0.0:
-                # Shouldn't happen
-                _log.warning(f'PPA price in Year 0 ({year_0_ppa_price}) is not zero, this is unexpected.')
-
-            # TODO (maybe)
-            # ppa_revenue_all_years = [
-            #     *([year_0_ppa_price] * (self._pre_revenue_years_count - 1)),
-            #     *ret[ppa_price_row_index][self._pre_revenue_years_count :],
-            # ]
-            # ret[_get_row_index('PPA price (cents/kWh)')][1:] = ppa_revenue_all_years
-
-            first_year_lppa_cents_per_kwh: float | str = 'NaN'
-            first_year_pv_annual_energy = ret[_get_row_index(pv_of_annual_energy_row_name)][1]
-
-            if (
-                isinstance(first_year_pv_annual_energy, int) or isinstance(first_year_pv_annual_energy, float)
-            ) and first_year_pv_annual_energy != 0.0:
-                # Note: expected to be same in all pre-revenue years since both price and revenue are zero until COD
-                first_year_lppa_cents_per_kwh = round(
-                    first_year_pv_of_ppa_revenue_usd * 100.0 / first_year_pv_annual_energy, 2
-                )
-
-            lppa_row_name = 'LPPA Levelized PPA price nominal (cents/kWh)'
-            ret[_get_row_index(lppa_row_name)][1:] = [
-                first_year_lppa_cents_per_kwh,
-                *([None] * self._pre_revenue_years_count),
-            ]
-
-        backfill_lppa_metrics()
-
-        def backfill_lcoh_nominal() -> None:
-            # WIP...
-
-            heat_provided_kwh_row_name = 'Heat provided (kWh)'
-            heat_provided_kwh_row_index = _get_row_index(
-                heat_provided_kwh_row_name, raise_exception_if_not_present=False
-            )
-            if heat_provided_kwh_row_index == -1:
-                return  # No heat provided row, nothing to do
-
-            heat_provided = cf_ret[heat_provided_kwh_row_index].copy()
-            heat_provided_backfilled = [0 if it == '' else (int(it) if is_int(it) else it) for it in heat_provided[1:]]
-
-            # TODO/WIP maybe duplicate heat provided row after after_tax_lcoe_and_ppa_price_header_row_title to mirror
-            #   electricity convention
-            # heat_provided_kwh_row_index = _get_row_index_after(
-            #     heat_provided_kwh_row_name, after_tax_lcoe_and_ppa_price_header_row_title
-            # )
-
-            ret[heat_provided_kwh_row_index][1:] = heat_provided_backfilled
-
-            # <Back>fill PV of heat provided
-            heat_provided_backfilled_pv_processed = heat_provided_backfilled.copy()
-            pv_of_heat_provided_backfilled_kwh = []
-            for year_ in range(self._pre_revenue_years_count):
-                pv_at_year_ = abs(
-                    round(
-                        npf.npv(
-                            self.nominal_discount_rate.quantity().to('dimensionless').magnitude,
-                            heat_provided_backfilled_pv_processed,
-                        )
-                    )
-                )
-
-                pv_of_heat_provided_backfilled_kwh.append(pv_at_year_)
-
-                heat_provided_at_year = heat_provided_backfilled_pv_processed.pop(0)
-                heat_provided_backfilled_pv_processed[0] = (
-                    heat_provided_backfilled_pv_processed[0] + heat_provided_at_year
-                )
-
-            # FIXME WIP TODO
-            # pv_of_annual_energy_row_name = 'Present value of annual energy nominal (kWh)'
-            # for pv_of_annual_energy_row_index in _get_row_indexes(pv_of_annual_energy_row_name):
-            #     ret[pv_of_annual_energy_row_index][1:] = [
-            #         pv_of_electricity_to_grid_backfilled_kwh[0],
-            #         *([''] * (self._pre_revenue_years_count - 1)),
-            #     ]
-
-            # pv_of_electricity_to_grid_backfilled_row_kwh = pv_of_electricity_to_grid_backfilled_kwh
-            pv_of_heat_provided_backfilled_row_kwh = pv_of_heat_provided_backfilled_kwh
-            pv_of_annual_costs_backfilled_row_values_usd = pv_of_annual_costs_backfilled_row[
-                1 if isinstance(pv_of_annual_costs_backfilled_row[0], str) else 0 :
-            ]
-
-            lcoh_nominal_backfilled = []
-            for _year in range(len(pv_of_annual_costs_backfilled_row_values_usd)):
-                entry: float | str = 'NaN'
-                if pv_of_heat_provided_backfilled_row_kwh[_year] != 0:
-                    entry = (
-                        pv_of_annual_costs_backfilled_row_values_usd[_year]
-                        * 100
-                        / pv_of_heat_provided_backfilled_row_kwh[_year]
-                    )
-
-                lcoh_nominal_backfilled.append(entry)
-
-            lcoh_nominal_row_name = 'LCOH Levelized cost of heating nominal ($/kWh)'  # FIXME WIP unit
-
-            # Insert new row if LCOE row does not exist (yet)
-            lcoh_nominal_row_index = _get_row_index(lcoh_nominal_row_name, raise_exception_if_not_present=False)
-
-            def _insert_row_before(before_row_name: str, row_name: str, row_content: list[Any] | None) -> None:
-                if row_content is None:
-                    row_content = ['' for _it in ret[[it[0] for it in ret].index(before_row_name)]][1:]
-
-                ret.insert(
-                    [it[0] for it in ret].index(before_row_name),
-                    [row_name, *row_content],
-                )
-
-            def _insert_blank_line_before(before_row_name: str) -> None:
-                _insert_row_before(
-                    before_row_name, '', ['' for _it in ret[[it[0] for it in ret].index(before_row_name)]][1:]
-                )
-
-            if lcoh_nominal_row_index == -1:
-                _insert_row_before('PROJECT STATE INCOME TAXES', lcoh_nominal_row_name, None)
-                _insert_blank_line_before('PROJECT STATE INCOME TAXES')
-                lcoh_nominal_row_index = [it[0] for it in ret].index(lcoh_nominal_row_name)
-
-            lcoh_nominal_backfilled_entry = lcoh_nominal_backfilled[0]
-            if isinstance(lcoh_nominal_backfilled_entry, float):
-                lcoh_nominal_backfilled_entry = round(lcoh_nominal_backfilled_entry, 2)
-
-            ret[lcoh_nominal_row_index][1:] = [
-                lcoh_nominal_backfilled_entry,
-                *([None] * (self._pre_revenue_years_count - 1)),
-            ]
-
-        backfill_lcoh_nominal()
-
-        return ret
-
-    @property
-    def sam_after_tax_net_cash_flow_all_years(self) -> list[float]:
-        return _after_tax_net_cash_flow_all_years(self.sam_cash_flow_profile, self._pre_revenue_years_count)
 
 
 def validate_read_parameters(model: Model) -> None:
@@ -813,46 +273,6 @@ def calculate_sam_economics(model: Model) -> SamEconomicsCalculations:
         sam_economics.capacity_payment_revenue_sources = _get_capacity_payment_revenue_sources(model)
 
     return sam_economics
-
-
-def _after_tax_net_cash_flow_all_years(cash_flow: list[list[Any]], pre_revenue_years_count: int) -> list[float]:
-    return _net_cash_flow_all_years(cash_flow, pre_revenue_years_count, tax_qualifier='after-tax')
-
-
-def _net_cash_flow_all_years(
-    cash_flow: list[list[Any]], pre_revenue_years_count: int, tax_qualifier='after-tax'
-) -> list[float]:
-    if tax_qualifier not in ['after-tax', 'pre-tax']:
-        raise ValueError(f'Invalid tax qualifier: {tax_qualifier}')
-
-    def _get_row(row_name__: str) -> list[Any]:
-        for r in cash_flow:
-            if r[0] == row_name__:
-                return r[1:]
-
-        raise ValueError(f'Could not find row with name {row_name__}')
-
-    def _construction_returns_row(_construction_tax_qualifier: str) -> list[Any]:
-        returns_row_name = (
-            f'Total {_construction_tax_qualifier} returns [construction] ($)'
-            if tax_qualifier == 'pre-tax'
-            else f'After-tax net cash flow [construction] ($)'
-        )
-        return _get_row(returns_row_name)
-
-    try:
-        construction_returns_row = _construction_returns_row(tax_qualifier)
-    except ValueError as ve:
-        if tax_qualifier == 'pre-tax':
-            # TODO log warning
-            construction_returns_row = _construction_returns_row('after-tax')
-        else:
-            raise ve
-
-    return [
-        *[float(it) for it in construction_returns_row if is_float(it)],
-        *[float(it) for it in _get_row(f'Total {tax_qualifier} returns ($)')[pre_revenue_years_count:] if is_float(it)],
-    ]
 
 
 def _get_project_npv_musd(single_owner: Singleowner, cash_flow: list[list[Any]], model: Model) -> float:
